refactor: move to hoist.ts
